requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
async-timeout==5.0.1
attrs==25.3.0
backports.zoneinfo==0.2.1
certifi==2026.2.25
charset-normalizer==3.4.5
dateparser==1.2.2
frozenlist==1.5.0
idna==3.11
multidict==6.1.0
propcache==0.2.0
pycryptodome==3.23.0
python-binance==1.0.35
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2026.1.post1
regex==2024.11.6
requests==2.32.4
six==1.17.0
typing-extensions==4.13.2
tzlocal==5.2
urllib3==2.2.3
websockets==13.1
yarl==1.15.2
